remove redundant cast

+++ src/app.cpp
@@ -304,7 +304,7 @@ int App::score(int x, int y, int seconds, bool gravity) const
     auto const nTiles = static_cast<double>(x * y);
     auto const tilesPerSec = nTiles / static_cast<double>(seconds);
 
-    auto const sizeBonus = std::sqrt(nTiles / static_cast<double>(14.0 * 6.0));
+    auto const sizeBonus = std::sqrt(nTiles / 14.0 * 6.0);
     auto const points = tilesPerSec / 0.14 * 100.0;
     auto const gravityBonus = gravity ? 2.0 : 1.0;
